refactor: update chart data handling and improve UI components feat: introduce ChartsCanvaWidget for enhanced chart rendering
Summary:
This commit refactors the chart data handling in the application and introduces a new `ChartsCanvaWidget` for improved chart rendering.

Key changes:
- **Chart Data Handling:**
    - Updated `ChartsInteractorImpl` to group training data by name and calculate average weights more efficiently.
    - Refactored `SingleChartDomainModel` and `SingleChartUiModel` to use `Float` instead of `Double` for values.
- **UI Components:**
    - Added `ChartsCanvaWidget` to encapsulate chart rendering logic.
    - Updated `ChartsWidget` to utilize `ChartsCanvaWidget` for displaying charts.
    - Adjusted padding and background settings in various UI components for better layout.

+++ feature/charts/src/test/kotlin/io/github/stslex/workeeper/feature/charts/ui/mvi/handler/PagingHandlerTest.kt
@@ -113,7 +113,7 @@ internal class PagingHandlerTest {
         val domainData = listOf(
             SingleChartDomainModel(
                 name = "Push ups",
-                values = listOf(1.0, 2.0, 3.0),
+                values = listOf(1.0f, 2.0f, 3.0f),
             ),
         )
 
@@ -286,7 +286,7 @@ internal class PagingHandlerTest {
     @Test
     fun `chart results mapper errors are handled gracefully`() = runTest {
         val domainData = listOf(
-            SingleChartDomainModel(name = "Test", values = listOf(1.0, 2.0)),
+            SingleChartDomainModel(name = "Test", values = listOf(1.0f, 2.0f)),
         )
 
         every { commonStore.homeSelectedStartDate } returns flowOf(null)
@@ -349,7 +349,7 @@ internal class PagingHandlerTest {
         val largeDataSet = (1..10).map { index ->
             SingleChartDomainModel(
                 name = "Exercise_$index",
-                values = (1..5).map { it.toDouble() },
+                values = (1..5).map { it.toFloat() },
             )
         }
 
@@ -502,9 +502,9 @@ internal class PagingHandlerTest {
             SingleChartDomainModel(name = "", values = emptyList()),
             SingleChartDomainModel(
                 name = "   ",
-                values = listOf(Double.NaN, Double.POSITIVE_INFINITY),
+                values = listOf(Float.NaN, Float.POSITIVE_INFINITY),
             ),
-            SingleChartDomainModel(name = "Valid", values = listOf(-1.0, 0.0, 1.0)),
+            SingleChartDomainModel(name = "Valid", values = listOf(-1.0f, 0.0f, 1.0f)),
         )
 
         every { commonStore.homeSelectedStartDate } returns flowOf(1500000L)
